fef4 and fef8 did not quite work properly: the exponent part is an int, not a short!

+++ mach/proto/fp/fef4.c
@@ -12,7 +12,7 @@
 #include	"FP_types.h"
 
 struct	fef4_returns {
-	short	e;
+	int	e;
 	_float	f;
 };
